refactor(#449): split enhancedSpeechUtils.ts into voiceSelector + speaker
